Logs out user after application installation
Ensures the user is logged out after the installation of applications
to apply the configuration changes and make the new applications available
immediately. This enhances the user experience by preventing the need for
a manual logout.

Adds `graphical.target` to the service's `after` list to ensure the graphical
interface is available before the service starts.